feat: add new file with question widget
DT-2076
